backend/requirements.txt
# backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pywebview>=4.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
GitPython>=3.1.0
pytest>=7.0.0
httpx>=0.24.0
citeproc-py>=0.9.0
